+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/test_storage_account_scenarios.py
@@ -10,7 +10,6 @@ from .storage_test_util import StorageScenarioMixin
 
 @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2016-12-01')
 class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
-
     @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2017-06-01')
     @ResourceGroupPreparer(name_prefix='cli_test_storage_service_endpoints')
     @StorageAccountPreparer()
@@ -120,6 +119,19 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
         self.cmd('storage account delete -g {} -n {} --yes'.format(resource_group, name))
         self.cmd('storage account check-name --name {}'.format(name),
                  checks=JMESPathCheck('nameAvailable', True))
+
+    @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2017-10-01')
+    @ResourceGroupPreparer(parameter_name_for_location='location', location='southcentralus')
+    def test_create_storage_account_v2(self, resource_group, location):
+        name = self.create_random_name(prefix='cli', length=24)
+
+        self.cmd('storage account create -n {} -g {} -l {} --kind StorageV2'.format(name, resource_group, location),
+                 checks=[JMESPathCheck('kind', 'StorageV2')])
+
+        self.cmd('storage account check-name --name {}'.format(name), checks=[
+            JMESPathCheck('nameAvailable', False),
+            JMESPathCheck('reason', 'AlreadyExists')
+        ])
 
     @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2016-01-01')
     @ResourceGroupPreparer(location='southcentralus')