+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.1.15
+++++++
+* Fix incorrect order of parameters for container logs
+
+0.1.14
+++++++
+* Fixed default ports regression
+
 0.1.13
 ++++++
 * minor fixes